Introduce disable_cpu_template in the runtime config
In this change we introduce DisableCPUTemplate in the runtime config.

By default disable_cpu_template param is set to false, in which case
CPUTemplate will be populated in the MachineConfiguration by either the
value in the runtime config or the firecracker-control client.

When the value is set to true, the CPUTemplate will simply be ignired in
MachineConfiguration and set to an empty string. This is particularly
useful on AMD (and possibly Arm) platform, since currently only Intel
x86 flavourd C3 and T2 templates are supported by the firecracker.

Signed-off-by: jhe <jingkai@hey.com>